Presidential Election 1988

The second presidential election of Sri Lanka was held in 1988. Its results are published as a pdf file. The data available in this file is scraped and summarised as a data-frame Election1988.

Usage

data(Election1988)

Format

A data frame with 6 columns and 1406 rows.

Year

Year of the presidential election

District

District Names

Electorate

Electorate Names, Postal District Votes and Final District Results

ColNames

Candidate Names, Total No of Valid Votes, No of Rejected Votes, Total No of Votes Polled and No of Registered Electors

Votes

No of votes for the ColNames variable

Percentage

Percentage of votes with regard to the No of Registered Electors and the Total No of Votes Polled for Total No of Votes Polled and all other ColNames variable, respectively

Source

Extracted from the pdf published by the Sri Lankan Elections commission at https://elections.gov.lk or the project that I have completed in scraping the data from the pdf file at : https://github.com/Amalan-ConStat/PresidentialElection

Details

There are 22 districts and 160 electorates in Sri Lanka. In 1988 three candidates participated in this election and Ranasinghe Premadasa became the second executive president.

Examples

Election1988[Election1988$District == "Colombo" & Election1988$Electorate == "Colombo-North",]
#> # A tibble: 7 × 6
#>    Year District Electorate    ColNames                         Votes Percentage
#>   <dbl> <chr>    <chr>         <chr>                            <dbl>      <dbl>
#> 1  1988 Colombo  Colombo-North Mr.Ranasinghe Premadasa          22389      63.3 
#> 2  1988 Colombo  Colombo-North Mrs.Sirimavo Ratwatte Dias Band… 10744      30.4 
#> 3  1988 Colombo  Colombo-North Mr.Ahangama Vithanage Oswin Nan…  2259       6.38
#> 4  1988 Colombo  Colombo-North Total No of Valid Votes          35392      98.2 
#> 5  1988 Colombo  Colombo-North No of Rejected Votes               647       1.8 
#> 6  1988 Colombo  Colombo-North Total No of Votes Polled         36039      66.6 
#> 7  1988 Colombo  Colombo-North No of Registered Electors        54149      NA