Presidential Election 2005

The fifth presidential election of Sri Lanka was held in 2005. Its results are published as a pdf file. The data available in this file is scraped and summarised as a data-frame Election2005.

Usage

data(Election2005)

Format

A data frame with 6 columns and 3494 rows.

Year

Year of the presidential election

District

District Names

Electorate

Electorate Names, Postal District Votes, Final District Results and Displaced District Votes

ColNames

Candidate Names, Total No of Valid Votes, No of Rejected Votes, Total No of Votes Polled and No of Registered Electors

Votes

No of votes for the ColNames variable

Percentage

Percentage of votes with regard to the No of Registered Electors and the Total No of Votes Polled for Total No of Votes Polled and all other ColNames variable, respectively

Source

Extracted from the pdf published by the Sri Lankan Elections commission at https://elections.gov.lk/web/wp-content/uploads/election-results/presidential-elections/PresidentialElections2005(allisland).pdf or the project that I have completed in scraping the data from the pdf file at : https://github.com/Amalan-ConStat/PresidentialElection

Details

There are 22 districts and 160 electorates in Sri Lanka. In 2005 thirteen candidates participated in this election and Mahinda Rajapaksha became the fourth executive president.

Examples

Election2005[Election2005$District == "Colombo" & Election2005$Electorate == "Colombo-North",]
#> # A tibble: 17 x 6
#>     Year District Electorate    ColNames                           Votes Perce~1
#>    <dbl> <chr>    <chr>         <chr>                              <dbl>   <dbl>
#>  1  2005 Colombo  Colombo-North Wimal Geeganage                       20    0.03
#>  2  2005 Colombo  Colombo-North Chamil Jayaneththi                    59    0.1 
#>  3  2005 Colombo  Colombo-North Ajith Kumara Jayaweera Arachchige     27    0.04
#>  4  2005 Colombo  Colombo-North Siritunga Jayasuriya                 270    0.45
#>  5  2005 Colombo  Colombo-North P. Nelson Perera                       5    0.01
#>  6  2005 Colombo  Colombo-North Wije Dias                             27    0.04
#>  7  2005 Colombo  Colombo-North Anura De Silva                        39    0.06
#>  8  2005 Colombo  Colombo-North Aruna de Soyza                        17    0.03
#>  9  2005 Colombo  Colombo-North Mahinda Rajapaksha                 14270   23.7 
#> 10  2005 Colombo  Colombo-North Ranil Wickramasinghe               45161   75.1 
#> 11  2005 Colombo  Colombo-North Achala Ashoka Suraweera              112    0.19
#> 12  2005 Colombo  Colombo-North Victor Hettigoda                     101    0.17
#> 13  2005 Colombo  Colombo-North Hewaheenipallage Shantha Dharmadw~     3    0   
#> 14  2005 Colombo  Colombo-North Total No of Valid Votes            60111   98.5 
#> 15  2005 Colombo  Colombo-North No of Rejected Votes                 939    1.54
#> 16  2005 Colombo  Colombo-North Total No of Votes Polled           61050   71.0 
#> 17  2005 Colombo  Colombo-North No of Registered Electors          85967   NA   
#> # ... with abbreviated variable name 1: Percentage